In this example, the EncapsulationExample class has two private members, id and name. The getId and getName methods are used to access the values of id and name, respectively, while the setId and setName methods are used to modify the values.
In the Main class, an instance of EncapsulationExample is created and the values of id and name are set using the setId and setName methods. The values of id and name are then displayed using the getId and getName methods.

By using encapsulation, the data is protected and can only be accessed and modified through the provided methods, ensuring that the data remains consistent and cannot be modified in an unintended way.